ASearch

+++ ASearchAlgorithm/ASearchalgorithm/src/main/java/entitys/ComplexEdge.java
@@ -5,27 +5,27 @@ import org.jgrapht.graph.DefaultEdge;
 
 public class ComplexEdge extends DefaultEdge {
 
-    private int reDistance = 0;
-    private int imDistance = 0;
+    private double reDistance = 0;
+    private double imDistance = 0;
 
-    public ComplexEdge(int imDistance, int reDistance) {
+    public ComplexEdge(double imDistance, double reDistance) {
         this.imDistance = imDistance;
         this.reDistance = reDistance;
     }
 
-    public int getReDistance() {
+    public double getReDistance() {
         return reDistance;
     }
 
-    public void setReDistance(int reDistance) {
+    public void setReDistance(double reDistance) {
         this.reDistance = reDistance;
     }
 
-    public int getImDistance() {
+    public double getImDistance() {
         return imDistance;
     }
 
-    public void setImDistance(int imDistance) {
+    public void setImDistance(double imDistance) {
         this.imDistance = imDistance;
     }
